Refactor hook initialization, enhance thread safety, and improve memory operations
- Added `std::mutex` to `Hooks` for thread safety during initialization and uninstallation.
- Introduced exception handling in `Hook_Present` for robustness.
- Updated `inject.py` with 64-bit compatible function prototypes and error handling for memory operations.
- Switched `m_cachedSlot` to use `UINT` for consistency.
- Changed recording output path in `CubemapManager` to a relative path.
- Display debug message on DLL attach for testing purposes.

+++ src/Camera/CameraController.cpp
@@ -8,7 +8,7 @@ namespace Camera {
 
     void CameraController::Reset() {
         m_cachedCameraBuffer = nullptr;
-        m_cachedSlot = -1;
+        m_cachedSlot = (UINT)-1;
     }
 
     void CameraController::SetTargetFace(CubeFace face) {
